feat: Delete rows from variation_hgvs related to deleted transcripts

+++ modules/Bio/EnsEMBL/Variation/Pipeline/GeneFactory.pm
@@ -115,7 +115,7 @@ sub fetch_input {
     } 
 
     # Remove Deleted transcripts
-    if (-e $self->param('update_diff')){
+    if (-e $self->param('update_diff') && @delete_transcripts){
         my $joined_ids = '"' . join('", "', @delete_transcripts) . '"';
         $dbc->do(qq{
                   DELETE FROM  transcript_variation
@@ -126,6 +126,24 @@ sub fetch_input {
                   DELETE FROM  MTMP_transcript_variation
                   WHERE   feature_stable_id IN ($joined_ids)
         }) if($mtmp);
+
+        $dbc->do(qq{
+                  DELETE WHOLE FROM variation_hgvs WHOLE
+                  INNER JOIN 
+                  (SELECT var_id.variation_id AS variation_id, SUBSTRING_INDEX(hgvs_transcript, ":", -1) AS hgvs_name
+                  FROM transcript_variation, (select variation_id from variation_feature WHERE variation_feature_id IN (
+                    SELECT DISTINCT(variation_feature_id) from transcript_variation WHERE feature_stable_id IN ($joined_ids) AND hgvs_transcript IS NOT NULL
+                    )) as var_id
+                  WHERE variation_feature_id = (SELECT variation_feature_id FROM variation_feature WHERE variation_id = var_id.variation_id) AND hgvs_transcript IS NOT NULL
+                  UNION ALL
+                  SELECT var_id.variation_id AS variation_id, SUBSTRING_INDEX(hgvs_protein, ":", -1) AS hgvs_name
+                  FROM transcript_variation, (select variation_id from variation_feature WHERE variation_feature_id IN (
+                    SELECT DISTINCT(variation_feature_id) from transcript_variation WHERE feature_stable_id IN ($joined_ids) AND hgvs_protein IS NOT NULL
+                    )) as var_id
+                  WHERE variation_feature_id = (SELECT variation_feature_id FROM variation_feature WHERE variation_id = var_id.variation_id) AND hgvs_protein IS NOT NULL) SUBSET
+                  ON WHOLE.variation_id=SUBSET.variation_id AND WHOLE.hgvs_name=SUBSET.hgvs_name;
+        });
+
     }
 
     $self->param('gene_output_ids', \@gene_output_ids);
